third_party/blink/web_tests/transforms/scale-parsing-expected.txt

This is a testharness.js-based test.
[FAIL] "scale: calc(1 * 2);" should parse as "2"
  assert_equals: expected "2" but got "calc(2)"
[FAIL] "scale: calc(1 * 2) calc(2 / 4);" should parse as "2 0.5"
  assert_equals: expected "2 0.5" but got "calc(2) calc(0.5)"
[FAIL] "scale: calc(1 * 2) calc(2 / 4) calc(0.5);" should parse as "2 0.5 0.5"
  assert_equals: expected "2 0.5 0.5" but got "calc(2) calc(0.5) calc(0.5)"
[FAIL] "scale: calc(1 * 2) calc(2 / 4) calc(0.5%);" should parse as "2 0.5 0.005"
  assert_equals: expected "2 0.5 0.005" but got "calc(2) calc(0.5) 0.005"
Harness: the test ran to completion.
